+++ examples/resources/algolia_ingestion_task/import.sh
@@ -1,3 +1,3 @@
 # Import an Ingestion task by its UUID. The `cursor` cannot be recovered on
 # import and is left null.
-terraform import algolia_ingestion_task.shopify_to_products 6c02aeb1-775e-418e-870b-1faccd4b2c0f
+terraform import algolia_ingestion_task.csv_to_products 6c02aeb1-775e-418e-870b-1faccd4b2c0f